VR: split GraphElements
